backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
Pillow==10.1.0
numpy==1.26.2
httpx==0.25.2
